refactor : Moved the chatbot package services to more suitable packages.
